+++ R/utils.R
@@ -278,14 +278,17 @@ strcat <- function(a, b) .Call(rnng_strcat, a, b)
 #' Configure Next Mode
 #'
 #' Configures send mode 'next' by registering functions for custom serialization
-#'     and unserialization of external pointer reference objects, allowing these
+#'     and unserialization of non-system reference objects, allowing these
 #'     to be sent and received between different R sessions.
 #'
 #' @param refhook \strong{either} a list or pairlist of two functions: the
-#'     signature for the first must accept a list of external pointer objects
-#'     and return a raw vector, e.g. \code{torch::torch_serialize}, and the
-#'     second must accept a raw vector and return a list of external pointer
-#'     objects, e.g. \code{torch::torch_load},\cr \strong{or else} NULL to reset.
+#'     signature for the first must accept a list of reference objects (which
+#'     inherit 'class') and return a raw vector, e.g.
+#'     \code{torch::torch_serialize}, and the second must accept a raw vector
+#'     and return a list of reference objects, e.g. \code{torch::torch_load},
+#'     \cr \strong{or else} NULL to reset.
+#' @param class [default 'torch_tensor'] a character string representing the
+#'     class of object that these serialization function will be applied to.
 #' @param mark [default FALSE] (for advanced use only) logical value, whether to
 #'     mark serialized data with a special bit.
 #'
@@ -304,5 +307,5 @@ strcat <- function(a, b) .Call(rnng_strcat, a, b)
 #'
 #' @export
 #'
-next_config <- function(refhook = list(), mark = FALSE)
-  .Call(rnng_next_config, refhook, mark)
+next_config <- function(refhook = list(), class = "torch_tensor", mark = FALSE)
+  .Call(rnng_next_config, refhook, class, mark)